+++ src/SpeckleToDM.cpp
@@ -93,7 +93,7 @@ void SpeckleToDM::generateMapFromSpeckle(const cv::Point2d kvecs, double amp, do
         for(c=0; c<DM_X_SIZE; c++){
             phy = (float)(((1.0/DM_Y_SIZE)*r-0.5)*kvecs.y);
             phx = (float)(((1.0/DM_X_SIZE)*c-0.5)*kvecs.x);
-            *(map + r*DM_X_SIZE + c) += (float)amp*std::cos(phx + phy + phase);
+            *(map + r*DM_X_SIZE + c) += (float)amp*cosf(phx + phy + phase);
 
         }
